backend/requirements.txt
# =====================================================
# celx-atendimento - Backend Dependencies
# =====================================================

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# AI - LangChain & LangGraph
langchain==0.1.17
langchain-core==0.1.52
langchain-community>=0.0.36

# OpenRouter (via LangChain)
langchain-openai==0.1.6
langchain-anthropic==0.1.3

# Redis
redis==5.0.1

# Celery
celery==5.3.6

# File Processing
python-docx==1.1.0
openpyxl==3.1.2
aiofiles==23.2.1

# Cryptography
cryptography==42.0.2
fernet==1.0.1

# HTTP Client
httpx==0.26.0

# Utils
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0  # For test client
